requirements.txt
# Clinic Scheduler Agent - Core Dependencies
# Install with: pip install -r requirements.txt --pre

# Environment & Utilities
python-dotenv>=1.0.0

# Microsoft Agent Framework (Preview)
agent-framework
agent-framework-azure-ai

# Azure Authentication & LLM Client
azure-identity
openai

# Web UI
gradio